refactored level editor file into smaller files
