+++ frontend/src/components/editor/editor-scene.tsx
@@ -1,34 +1,24 @@
 "use client";
 
-import {
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-} from "react";
-import { Canvas, useThree } from "@react-three/fiber";
+import { useCallback, useMemo, useRef } from "react";
+import { Canvas } from "@react-three/fiber";
 import { OrbitControls } from "@react-three/drei";
 import * as THREE from "three";
 import { BLOCK_REGISTRY } from "@/lib/block-registry";
 import type { LevelData, LevelDimensions } from "@/types/level";
 import { getLevelDimensions } from "@/types/level";
 import type { EditorTool, EditorMode } from "./tool-panel";
+import type { HoverCell, SceneAPI } from "./scene/types";
+import { CameraSetup, SceneController } from "./scene/scene-controller";
+import {
+  BoundsBox,
+  GridLines,
+  HoverPreviewContent,
+  ResizePreview,
+  SpawnMarker,
+} from "./scene/scene-objects";
 
-export type HoverCell = { x: number; y: number; z: number };
-
-export type SceneAPI = {
-  // Step current layer by `dir` units. Clamps internally to [0, layerCount-1]
-  // and returns the actual new layer after clamping.
-  applyLayerStep: (dir: number) => number;
-  // Set current layer to absolute value. `layerCountOverride` lets callers
-  // pass the new layer count when the React state level hasn't propagated
-  // yet (e.g. immediately after add-layer or resize).
-  applyLayerSet: (target: number, layerCountOverride?: number) => number;
-  // Perform the current tool's action at the current hover cell. Used by
-  // the editor's space-key handler to paint the initial cell on press;
-  // hold-to-paint on subsequent hover changes is handled internally.
-  performActionAtHover: () => void;
-};
+export type { HoverCell, SceneAPI } from "./scene/types";
 
 interface EditorSceneProps {
   level: LevelData;
@@ -49,29 +39,6 @@ interface EditorSceneProps {
   onErase: (x: number, y: number, z: number) => void;
   onSetSpawn: (x: number, y: number, z: number) => void;
   onLayerChange?: (newLayer: number) => void;
-}
-
-const HINT_COLORS: Record<string, string> = {
-  air: "#7dd3fc",
-  empty: "#71717a",
-};
-
-function buildBlockList(
-  level: LevelData,
-): Array<{ x: number; y: number; z: number }> {
-  const out: Array<{ x: number; y: number; z: number }> = [];
-  for (const layerName in level.layers) {
-    const layerIndex = parseInt(layerName.split("_")[1]) || 0;
-    const layer = level.layers[layerName];
-    for (let z = 0; z < layer.length; z++) {
-      const row = layer[z];
-      for (let x = 0; x < row.length; x++) {
-        if (row[x] === "air") continue;
-        out.push({ x, y: layerIndex, z });
-      }
-    }
-  }
-  return out;
 }
 
 export function EditorScene({
@@ -279,533 +246,5 @@ export function EditorScene({
         maxPolarAngle={Math.PI / 2 - 0.05}
       />
     </Canvas>
-  );
-}
-
-function CameraSetup({ target }: { target: [number, number, number] }) {
-  const camera = useThree((s) => s.camera);
-  useEffect(() => {
-    camera.lookAt(target[0], target[1], target[2]);
-    camera.updateProjectionMatrix();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-  return null;
-}
-
-// Owns DOM pointer listeners, the raycast plane, last-cursor state, and
-// the imperative API for layer shifts. Lives inside <Canvas> so it can
-// use useThree.
-function SceneController({
-  initialActiveLayer,
-  sceneApiRef,
-  controlsRef,
-  layerGroupRef,
-  hoverGroupRef,
-  planeRef,
-  tool,
-  selectedBlockId,
-  mode,
-  level,
-  width,
-  depth,
-  spaceHeldRef,
-  onPaint,
-  onErase,
-  onSetSpawn,
-  onHoverChange,
-  onLayerChange,
-}: {
-  initialActiveLayer: number;
-  sceneApiRef?: React.RefObject<SceneAPI | null>;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  controlsRef?: React.RefObject<any>;
-  layerGroupRef: React.RefObject<THREE.Group | null>;
-  hoverGroupRef: React.RefObject<THREE.Group | null>;
-  planeRef: React.RefObject<THREE.Plane>;
-  tool: EditorTool;
-  selectedBlockId: string;
-  mode: EditorMode;
-  level: LevelData;
-  width: number;
-  depth: number;
-  spaceHeldRef?: React.RefObject<boolean>;
-  onPaint: (x: number, y: number, z: number, blockId: string) => void;
-  onErase: (x: number, y: number, z: number) => void;
-  onSetSpawn: (x: number, y: number, z: number) => void;
-  onHoverChange: (cell: HoverCell | null) => void;
-  onLayerChange?: (newLayer: number) => void;
-}) {
-  const { camera, gl, raycaster } = useThree();
-
-  const lastCursorRef = useRef<{ x: number; y: number } | null>(null);
-  const blocks = useMemo(() => buildBlockList(level), [level]);
-
-  // Source of truth for which layer the scene is on. Editor page must
-  // route ALL layer changes through applyLayerStep/applyLayerSet so
-  // camera + group + plane + state can't drift apart.
-  const currentLayerRef = useRef(initialActiveLayer);
-
-  // Throttles hold-space-to-paint so repeated pointermove events within
-  // the same cell don't spam the undo history.
-  const lastActionCellRef = useRef<HoverCell | null>(null);
-
-  const stateRef = useRef({
-    tool,
-    selectedBlockId,
-    mode,
-    width,
-    depth,
-    blocks,
-    level,
-    spaceHeldRef,
-    onPaint,
-    onErase,
-    onSetSpawn,
-    onHoverChange,
-    onLayerChange,
-  });
-  stateRef.current = {
-    tool,
-    selectedBlockId,
-    mode,
-    width,
-    depth,
-    blocks,
-    level,
-    spaceHeldRef,
-    onPaint,
-    onErase,
-    onSetSpawn,
-    onHoverChange,
-    onLayerChange,
-  };
-
-  // Reused temp objects — avoid per-frame allocation.
-  const tmps = useMemo(
-    () => ({
-      box: new THREE.Box3(),
-      min: new THREE.Vector3(),
-      max: new THREE.Vector3(),
-      hit: new THREE.Vector3(),
-      intersection: new THREE.Vector3(),
-      ndc: new THREE.Vector2(),
-    }),
-    [],
-  );
-
-  const computeHover = useCallback(
-    (cx: number, cy: number): HoverCell | null => {
-      const s = stateRef.current;
-      const rect = gl.domElement.getBoundingClientRect();
-      tmps.ndc.x = ((cx - rect.left) / rect.width) * 2 - 1;
-      tmps.ndc.y = -((cy - rect.top) / rect.height) * 2 + 1;
-      raycaster.setFromCamera(tmps.ndc, camera);
-
-      if (s.tool === "erase") {
-        let best: HoverCell | null = null;
-        let bestDistSq = Infinity;
-        for (let i = 0; i < s.blocks.length; i++) {
-          const b = s.blocks[i];
-          tmps.min.set(b.x - 0.5, b.y - 0.5, b.z - 0.5);
-          tmps.max.set(b.x + 0.5, b.y + 0.5, b.z + 0.5);
-          tmps.box.set(tmps.min, tmps.max);
-          if (!raycaster.ray.intersectBox(tmps.box, tmps.hit)) continue;
-          const dx = tmps.hit.x - camera.position.x;
-          const dy = tmps.hit.y - camera.position.y;
-          const dz = tmps.hit.z - camera.position.z;
-          const distSq = dx * dx + dy * dy + dz * dz;
-          if (distSq < bestDistSq) {
-            bestDistSq = distSq;
-            best = { x: b.x, y: b.y, z: b.z };
-          }
-        }
-        return best;
-      }
-
-      if (!raycaster.ray.intersectPlane(planeRef.current, tmps.intersection)) {
-        return null;
-      }
-      const ix = Math.floor(tmps.intersection.x + 0.5);
-      const iz = Math.floor(tmps.intersection.z + 0.5);
-      if (ix < 0 || ix >= s.width || iz < 0 || iz >= s.depth) return null;
-      const layerY = -planeRef.current.constant;
-      return { x: ix, y: layerY + 0.5, z: iz };
-    },
-    [gl, camera, raycaster, planeRef, tmps],
-  );
-
-  const performActionAtCell = useCallback((cell: HoverCell) => {
-    const s = stateRef.current;
-    if (s.tool === "paint") {
-      s.onPaint(cell.x, cell.y, cell.z, s.selectedBlockId);
-    } else if (s.tool === "erase") {
-      s.onErase(cell.x, cell.y, cell.z);
-    } else if (s.tool === "spawn") {
-      s.onSetSpawn(cell.x, cell.y, cell.z);
-    }
-    lastActionCellRef.current = cell;
-  }, []);
-
-  const updateHover = useCallback(() => {
-    const last = lastCursorRef.current;
-    const s = stateRef.current;
-    if (!last || s.mode !== "placement") {
-      if (hoverGroupRef.current) hoverGroupRef.current.visible = false;
-      s.onHoverChange(null);
-      return;
-    }
-    const cell = computeHover(last.x, last.y);
-    if (cell && hoverGroupRef.current) {
-      hoverGroupRef.current.position.set(cell.x, cell.y, cell.z);
-      hoverGroupRef.current.visible = true;
-    } else if (hoverGroupRef.current) {
-      hoverGroupRef.current.visible = false;
-    }
-    s.onHoverChange(cell);
-
-    // Hold-space-to-paint: re-fire the tool action whenever the hover
-    // moves to a different cell while space is held.
-    if (cell && s.spaceHeldRef?.current) {
-      const last = lastActionCellRef.current;
-      if (
-        !last ||
-        last.x !== cell.x ||
-        last.y !== cell.y ||
-        last.z !== cell.z
-      ) {
-        performActionAtCell(cell);
-      }
-    }
-  }, [computeHover, hoverGroupRef, performActionAtCell]);
-
-  // DOM pointer listeners — attached once.
-  useEffect(() => {
-    const canvas = gl.domElement;
-    const handleMove = (e: PointerEvent) => {
-      lastCursorRef.current = { x: e.clientX, y: e.clientY };
-      updateHover();
-    };
-    const handleDown = (e: PointerEvent) => {
-      if (e.button !== 0) return;
-      const s = stateRef.current;
-      if (s.mode !== "placement") return;
-      const cell = computeHover(e.clientX, e.clientY);
-      if (!cell) return;
-      performActionAtCell(cell);
-    };
-    const handleLeave = () => {
-      lastCursorRef.current = null;
-      if (hoverGroupRef.current) hoverGroupRef.current.visible = false;
-      stateRef.current.onHoverChange(null);
-    };
-    canvas.addEventListener("pointermove", handleMove);
-    canvas.addEventListener("pointerdown", handleDown);
-    canvas.addEventListener("pointerleave", handleLeave);
-    return () => {
-      canvas.removeEventListener("pointermove", handleMove);
-      canvas.removeEventListener("pointerdown", handleDown);
-      canvas.removeEventListener("pointerleave", handleLeave);
-    };
-  }, [gl, computeHover, updateHover, performActionAtCell, hoverGroupRef]);
-
-  // Re-raycast on tool/mode change so the preview reflects the new state
-  // without requiring a cursor move.
-  useEffect(() => {
-    updateHover();
-  }, [tool, mode, updateHover]);
-
-  // Single internal mutator. All public APIs route through this so bounds
-  // checking lives in exactly one place and the visual state can never
-  // disagree with currentLayerRef.
-  const shiftToLayer = useCallback(
-    (target: number, layerCountOverride?: number): number => {
-      const s = stateRef.current;
-      const layerCount =
-        layerCountOverride ?? Object.keys(s.level.layers).length;
-      const maxLayer = Math.max(layerCount - 1, 0);
-      const clamped = Math.min(Math.max(target, 0), maxLayer);
-      const current = currentLayerRef.current;
-      if (clamped === current) return current;
-      const delta = clamped - current;
-      currentLayerRef.current = clamped;
-
-      if (controlsRef?.current) {
-        controlsRef.current.target.y += delta;
-        controlsRef.current.object.position.y += delta;
-        controlsRef.current.update();
-      }
-      if (layerGroupRef.current) {
-        layerGroupRef.current.position.y += delta;
-      }
-      // Plane: ax+by+cz+d=0 with normal (0,1,0) → y = -d. Shifting up by
-      // delta means y_new = y_old + delta → d_new = d_old - delta.
-      planeRef.current.constant -= delta;
-      updateHover();
-      s.onLayerChange?.(clamped);
-      return clamped;
-    },
-    [controlsRef, layerGroupRef, planeRef, updateHover],
-  );
-
-  // Expose imperative scene API.
-  useEffect(() => {
-    if (!sceneApiRef) return;
-    const api: SceneAPI = {
-      applyLayerStep: (dir: number) =>
-        shiftToLayer(currentLayerRef.current + dir),
-      applyLayerSet: (target: number, layerCountOverride?: number) =>
-        shiftToLayer(target, layerCountOverride),
-      performActionAtHover: () => {
-        const s = stateRef.current;
-        if (s.mode !== "placement") return;
-        const last = lastCursorRef.current;
-        if (!last) return;
-        const cell = computeHover(last.x, last.y);
-        if (!cell) return;
-        performActionAtCell(cell);
-      },
-    };
-    sceneApiRef.current = api;
-    return () => {
-      if (sceneApiRef.current === api) sceneApiRef.current = null;
-    };
-  }, [sceneApiRef, shiftToLayer, computeHover, performActionAtCell]);
-
-  return null;
-}
-
-function HoverPreviewContent({
-  tool,
-  blockId,
-}: {
-  tool: EditorTool;
-  blockId: string;
-}) {
-  if (tool === "spawn") {
-    return (
-      <mesh position={[0, 0.5, 0]}>
-        <coneGeometry args={[0.3, 0.8, 8]} />
-        <meshBasicMaterial
-          color="#22c55e"
-          transparent
-          opacity={0.5}
-          depthWrite={false}
-        />
-      </mesh>
-    );
-  }
-  if (tool === "erase") {
-    return (
-      <mesh>
-        <boxGeometry args={[1.05, 1.05, 1.05]} />
-        <meshBasicMaterial
-          color="#ef4444"
-          transparent
-          opacity={0.35}
-          depthWrite={false}
-        />
-      </mesh>
-    );
-  }
-  const def = BLOCK_REGISTRY[blockId];
-  if (blockId === "air" || blockId === "empty" || !def) {
-    return (
-      <mesh>
-        <boxGeometry args={[0.95, 0.95, 0.95]} />
-        <meshBasicMaterial
-          color={HINT_COLORS[blockId] ?? "#9ca3af"}
-          transparent
-          opacity={0.5}
-          depthWrite={false}
-        />
-      </mesh>
-    );
-  }
-  const Component = def.component;
-  return (
-    <>
-      <Component position={[0, 0, 0]} blockDef={def} />
-      <mesh>
-        <boxGeometry args={[1.02, 1.02, 1.02]} />
-        <meshBasicMaterial
-          color="white"
-          transparent
-          opacity={0.4}
-          depthWrite={false}
-        />
-      </mesh>
-    </>
-  );
-}
-
-function GridLines({ width, depth }: { width: number; depth: number }) {
-  const geometry = useMemo(() => {
-    const pts: number[] = [];
-    for (let i = 0; i <= width; i++) {
-      const xPos = i - 0.5;
-      pts.push(xPos, 0, -0.5);
-      pts.push(xPos, 0, depth - 0.5);
-    }
-    for (let i = 0; i <= depth; i++) {
-      const zPos = i - 0.5;
-      pts.push(-0.5, 0, zPos);
-      pts.push(width - 0.5, 0, zPos);
-    }
-    const geom = new THREE.BufferGeometry();
-    geom.setAttribute("position", new THREE.Float32BufferAttribute(pts, 3));
-    return geom;
-  }, [width, depth]);
-
-  return (
-    <lineSegments geometry={geometry}>
-      <lineBasicMaterial color="#60a5fa" />
-    </lineSegments>
-  );
-}
-
-// Orientation 0=N(-Z), 1=E(+X), 2=S(+Z), 3=W(-X). The arrow geometry below
-// is built pointing toward -Z, so rotate around Y by (orientation * -π/2).
-const ARROW_GEOMETRY = (() => {
-  const g = new THREE.BufferGeometry();
-  g.setAttribute(
-    "position",
-    new THREE.Float32BufferAttribute(
-      [0, 0, -0.5, -0.3, 0, 0.3, 0.3, 0, 0.3],
-      3,
-    ),
-  );
-  g.computeVertexNormals();
-  return g;
-})();
-
-function SpawnMarker({
-  position,
-  orientation,
-}: {
-  position: [number, number, number];
-  orientation: number;
-}) {
-  const yaw = -(orientation * Math.PI) / 2;
-  return (
-    <group position={position}>
-      <mesh position={[0, 0.5, 0]}>
-        <coneGeometry args={[0.3, 0.8, 8]} />
-        <meshStandardMaterial
-          color="#22c55e"
-          emissive="#22c55e"
-          emissiveIntensity={0.3}
-        />
-      </mesh>
-      <mesh
-        position={[0, 0.05, 0]}
-        rotation={[0, yaw, 0]}
-        geometry={ARROW_GEOMETRY}
-      >
-        <meshBasicMaterial color="#22c55e" side={THREE.DoubleSide} />
-      </mesh>
-    </group>
-  );
-}
-
-function BoundsBox({
-  width,
-  height,
-  depth,
-}: {
-  width: number;
-  height: number;
-  depth: number;
-}) {
-  const cx = (width - 1) / 2;
-  const cy = (height - 1) / 2;
-  const cz = (depth - 1) / 2;
-  return (
-    <lineSegments position={[cx, cy, cz]}>
-      <edgesGeometry args={[new THREE.BoxGeometry(width, height, depth)]} />
-      <lineBasicMaterial color="#94a3b8" />
-    </lineSegments>
-  );
-}
-
-function ResizePreview({
-  current,
-  pending,
-}: {
-  current: LevelDimensions;
-  pending: LevelDimensions;
-}) {
-  const slabs: {
-    center: [number, number, number];
-    size: [number, number, number];
-  }[] = [];
-
-  if (pending.width < current.width) {
-    const dx = current.width - pending.width;
-    slabs.push({
-      center: [
-        (pending.width + current.width) / 2 - 0.5,
-        (current.height - 1) / 2,
-        (current.depth - 1) / 2,
-      ],
-      size: [dx, current.height, current.depth],
-    });
-  }
-  if (pending.height < current.height) {
-    const dy = current.height - pending.height;
-    const w = Math.min(pending.width, current.width);
-    const d = Math.min(pending.depth, current.depth);
-    slabs.push({
-      center: [
-        (w - 1) / 2,
-        (pending.height + current.height) / 2 - 0.5,
-        (d - 1) / 2,
-      ],
-      size: [w, dy, d],
-    });
-  }
-  if (pending.depth < current.depth) {
-    const dz = current.depth - pending.depth;
-    const w = Math.min(pending.width, current.width);
-    const h = Math.min(pending.height, current.height);
-    slabs.push({
-      center: [
-        (w - 1) / 2,
-        (h - 1) / 2,
-        (pending.depth + current.depth) / 2 - 0.5,
-      ],
-      size: [w, h, dz],
-    });
-  }
-
-  const newCx = (pending.width - 1) / 2;
-  const newCy = (pending.height - 1) / 2;
-  const newCz = (pending.depth - 1) / 2;
-
-  return (
-    <group>
-      {slabs.map((slab, i) => (
-        <mesh key={i} position={slab.center}>
-          <boxGeometry args={slab.size} />
-          <meshBasicMaterial
-            color="#ef4444"
-            transparent
-            opacity={0.25}
-            depthWrite={false}
-          />
-        </mesh>
-      ))}
-      <lineSegments position={[newCx, newCy, newCz]}>
-        <edgesGeometry
-          args={[
-            new THREE.BoxGeometry(
-              pending.width,
-              pending.height,
-              pending.depth,
-            ),
-          ]}
-        />
-        <lineBasicMaterial color="#22c55e" />
-      </lineSegments>
-    </group>
   );
 }
